Refactors our new API code into seperate module
