+++ pyunifiprotect/__init__.py
@@ -1,6 +1,7 @@
 """Python Wrapper for Unifi Protect."""
+from pyunifiprotect.api import ProtectApiClient
 from pyunifiprotect.exceptions import Invalid, NotAuthorized, NvrError
-from pyunifiprotect.unifi_protect_server import ProtectApiClient, UpvServer
+from pyunifiprotect.unifi_protect_server import UpvServer
 
 __all__ = [
     "Invalid",
