+++ pyunifiprotect/cli.py
@@ -8,9 +8,10 @@ from typing import Any, Optional, Union
 
 import typer
 
+from pyunifiprotect.api import _LOGGER, ProtectApiClient
 from pyunifiprotect.data import WSPacket
 from pyunifiprotect.test_util import SampleDataGenerator
-from pyunifiprotect.unifi_protect_server import _LOGGER, ProtectApiClient, UpvServer
+from pyunifiprotect.unifi_protect_server import UpvServer
 
 try:
     from IPython import embed  # type: ignore
