+++ tests/test_unifi_protect_server.py
@@ -1,206 +1,182 @@
 """Tests for pyunifiprotect.unifi_protect_server."""
-# pylint: disable=pointless-statement
-# pylint: disable=protected-access
 
+import asyncio
 from datetime import datetime, timedelta
-import logging
-import time
+from io import BytesIO
+from typing import Any, Dict, List, Optional
 from unittest.mock import patch
 
+from PIL import Image
 import pytest
 
-from pyunifiprotect import ProtectApiClient
-from pyunifiprotect.data import EventType, ModelType
-from pyunifiprotect.unifi_protect_server import NEVER_RAN, WEBSOCKET_CHECK_INTERVAL
+from pyunifiprotect import UpvServer
+from pyunifiprotect.unifi_data import EventType, ModelType
 from pyunifiprotect.utils import to_js_time
-from tests.conftest import MockDatetime, compare_objs
+from tests.conftest import MockDatetime, MockWebsocket
 from tests.sample_data.constants import CONSTANTS
 
 
 @pytest.mark.asyncio
-async def test_api_client_creation():
+async def test_upvserver_creation():
     """Test we can create the object."""
 
-    client = ProtectApiClient("127.0.0.1", 0, "username", "password")
-    assert client
+    upv = UpvServer(None, "127.0.0.1", 0, "username", "password")
+    assert upv
 
 
 @pytest.mark.asyncio
-async def test_bootstrap(protect_client: ProtectApiClient):
-    """Verifies lookup of all object via ID"""
+async def test_websocket(old_protect_client: UpvServer, ws_messages: Dict[str, Dict[str, Any]]):
+    # wait for ws connection
+    for _ in range(60):
+        if old_protect_client.ws_connection is not None:
+            break
+        await asyncio.sleep(0.5)
 
-    assert protect_client.bootstrap.auth_user
+    ws_connect: Optional[MockWebsocket] = old_protect_client.ws_connection  # type: ignore
+    assert ws_connect is not None
 
-    for light in protect_client.bootstrap.lights.values():
-        light.camera
-        light.last_motion_event
+    while old_protect_client.ws_connection is not None:
+        await asyncio.sleep(0.1)
 
-    for camera in protect_client.bootstrap.cameras.values():
-        camera.last_motion_event
-        camera.last_ring_event
-        camera.last_smart_detect_event
-
-    for viewer in protect_client.bootstrap.viewers.values():
-        assert viewer.liveview
-
-    for liveview in protect_client.bootstrap.liveviews.values():
-        liveview.owner
-
-        for slot in liveview.slots:
-            assert len(slot.camera_ids) == len(slot.cameras)
-
-    for user in protect_client.bootstrap.users.values():
-        user.groups
-
-        if user.cloud_account is not None:
-            assert user.cloud_account.user == user
-
-    for event in protect_client.bootstrap.events.values():
-        event.smart_detect_events
+    assert ws_connect.count == len(ws_messages)
+    assert ws_connect.now == float(list(ws_messages.keys())[-1])
 
 
 @pytest.mark.asyncio
-async def test_bootstrap_construct(protect_client_no_debug: ProtectApiClient):
-    """Verifies lookup of all object via ID"""
+async def test_server_info(old_protect_client: UpvServer):
+    data = await old_protect_client.server_information()
 
-    protect_client = protect_client_no_debug
-    assert protect_client.bootstrap.auth_user
-
-    for light in protect_client.bootstrap.lights.values():
-        light.camera
-        light.last_motion_event
-
-    for camera in protect_client.bootstrap.cameras.values():
-        camera.last_motion_event
-        camera.last_ring_event
-        camera.last_smart_detect_event
-
-    for viewer in protect_client.bootstrap.viewers.values():
-        assert viewer.liveview
-
-    for liveview in protect_client.bootstrap.liveviews.values():
-        liveview.owner
-
-        for slot in liveview.slots:
-            assert len(slot.camera_ids) == len(slot.cameras)
-
-    for user in protect_client.bootstrap.users.values():
-        user.groups
-
-        if user.cloud_account is not None:
-            assert user.cloud_account.user == user
-
-    for event in protect_client.bootstrap.events.values():
-        event.smart_detect_events
+    old_protect_client.api_request.assert_called_with(  # type: ignore
+        url="bootstrap",
+        method="get",
+        require_auth=True,
+        raise_exception=True,
+    )
+    assert data == {
+        "server_id": CONSTANTS["server_id"],
+        "server_model": CONSTANTS["server_model"],
+        "server_name": CONSTANTS["server_name"],
+        "server_version": CONSTANTS["server_version"],
+        "server_ip": CONSTANTS["server_ip"],
+    }
 
 
 @pytest.mark.asyncio
 @patch("pyunifiprotect.unifi_protect_server.datetime", MockDatetime)
-async def test_get_events_raw_default(protect_client: ProtectApiClient, now: datetime):
-    events = await protect_client.get_events_raw()
+async def test_get_raw_events(old_protect_client: UpvServer, now: datetime):
+    events = await old_protect_client.get_raw_events()
 
-    end = now + timedelta(seconds=10)
-
-    protect_client.api_request.assert_called_with(  # type: ignore
+    old_protect_client.api_request.assert_called_with(  # type: ignore
         url="events",
         method="get",
         require_auth=True,
         raise_exception=True,
         params={
-            "start": to_js_time(end - timedelta(hours=24)),
-            "end": to_js_time(end),
+            "end": str(to_js_time(now + timedelta(seconds=10))),
+            "start": str(to_js_time(now - timedelta(seconds=86400))),
         },
     )
     assert len(events) == CONSTANTS["event_count"]
     for event in events:
-        assert event["type"] in EventType.values()
-        assert event["modelKey"] in ModelType.values()
+        event_type: str = event["type"]
+        model_type: str = event["modelKey"]
+
+        assert event_type in EventType.values()
+        assert model_type in ModelType.values()
 
 
 @pytest.mark.asyncio
-async def test_get_events_raw_limit(protect_client: ProtectApiClient):
-    await protect_client.get_events_raw(limit=10)
+async def test_raw_devices(old_protect_client: UpvServer):
+    all_model_types = [e.value for e in ModelType]
 
-    protect_client.api_request.assert_called_with(  # type: ignore
-        url="events",
+    data = await old_protect_client.get_raw_device_info()
+
+    old_protect_client.api_request.assert_called_with(  # type: ignore
+        url="bootstrap",
         method="get",
         require_auth=True,
         raise_exception=True,
-        params={"limit": 10},
     )
+    assert data.pop("authUserId") == CONSTANTS["user_id"]
+    assert data.pop("lastUpdateId") == CONSTANTS["last_update_id"]
+    data.pop("accessKey")
+    data.pop("legacyUFVs")
+    data.pop("nvr")
+    for key in data.keys():
+        model_type = key[:-1]
+        assert model_type in all_model_types
+        assert len(data[key]) == CONSTANTS["counts"][model_type]
+        for item in data[key]:
+            assert item["modelKey"] == model_type
+
+
+@pytest.mark.skipif(CONSTANTS.get("camera_thumbnail") is None, reason="No Camera thumbnail in test data")
+@pytest.mark.asyncio
+async def test_get_thumbnail(old_protect_client: UpvServer, camera):
+    data = await old_protect_client.get_thumbnail(camera_id=camera["id"])
+
+    old_protect_client.api_request_raw.assert_called_with(  # type: ignore
+        f"thumbnails/{CONSTANTS['camera_thumbnail']}",
+        params={
+            "h": "360.0",
+            "w": "640",
+        },
+    )
+
+    assert data is not None
+    img = Image.open(BytesIO(data))
+    assert img.width in (360, 640)  # Unifi may give a square image even though we did not request one
+    assert img.height == 360
 
 
 @pytest.mark.asyncio
-async def test_get_events_raw_cameras(protect_client: ProtectApiClient):
-    await protect_client.get_events_raw(limit=10, camera_ids=["test1", "test2"])
+@pytest.mark.skipif(not CONSTANTS.get("camera_online"), reason="No online camera in test data")
+@patch("pyunifiprotect.unifi_protect_server.datetime", MockDatetime)
+async def test_get_snapshot(old_protect_client: UpvServer, now, camera):
+    data = await old_protect_client.get_snapshot_image(camera_id=camera["id"])
+    if data is None:
+        return
 
-    protect_client.api_request.assert_called_with(  # type: ignore
-        url="events",
+    height = old_protect_client.devices[camera["id"]].get("image_height")
+    width = old_protect_client.devices[camera["id"]].get("image_width")
+
+    old_protect_client.api_request_raw.assert_called_with(  # type: ignore
+        f"cameras/{camera['id']}/snapshot",
+        params={
+            "h": height,
+            "ts": str(to_js_time(now)),
+            "force": "true",
+            "w": width,
+        },
+        raise_exception=False,
+    )
+
+    img = Image.open(BytesIO(data))
+    assert img.width == width
+    assert img.height == height
+
+
+@pytest.mark.asyncio
+async def test_get_live_views(old_protect_client: UpvServer, liveviews: List[Dict[str, Any]]):
+    data = await old_protect_client.get_live_views()
+
+    old_protect_client.api_request.assert_called_with(  # type: ignore
+        url="liveviews",
         method="get",
         require_auth=True,
         raise_exception=True,
-        params={"limit": 10, "cameras": "test1,test2"},
     )
 
+    assert len(data) == CONSTANTS["counts"]["liveview"]
+    for view in data:
+        mock_view = None
 
-@pytest.mark.asyncio
-async def test_get_events(protect_client: ProtectApiClient, raw_events):
-    expected_events = []
-    for event in raw_events:
-        if event["score"] >= 50 and event["type"] in EventType.device_events():
-            expected_events.append(event)
+        for item in liveviews:
+            if item.get("id") == view.get("id"):
+                mock_view = item
+                break
 
-    protect_client._minimum_score = 50
-
-    events = await protect_client.get_events()
-
-    assert len(events) == len(expected_events)
-    for index, event in enumerate(events):
-        compare_objs(event.model.value, expected_events[index], event.unifi_dict())
-
-
-@pytest.mark.asyncio
-async def test_check_ws_initial(protect_client: ProtectApiClient, caplog: pytest.LogCaptureFixture):
-    caplog.set_level(logging.DEBUG)
-
-    protect_client._last_websocket_check = NEVER_RAN
-    protect_client.disconnect_ws()
-    protect_client.ws_connection = None
-
-    active_ws = await protect_client.check_ws()
-
-    assert active_ws is True
-    assert ["Checking websocket"] == [rec.message for rec in caplog.records]
-
-
-@pytest.mark.asyncio
-async def test_check_ws_no_ws(protect_client: ProtectApiClient, caplog: pytest.LogCaptureFixture):
-    caplog.set_level(logging.DEBUG)
-
-    protect_client._last_websocket_check = time.monotonic()
-    protect_client.disconnect_ws()
-    protect_client.ws_connection = None
-
-    active_ws = await protect_client.check_ws()
-
-    assert active_ws is False
-
-    expected_logs = ["Unifi OS: Websocket connection not active, failing back to polling"]
-    assert expected_logs == [rec.message for rec in caplog.records]
-    assert caplog.records[0].levelname == "DEBUG"
-
-
-@pytest.mark.asyncio
-async def test_check_ws_reconnect(protect_client: ProtectApiClient, caplog: pytest.LogCaptureFixture):
-    caplog.set_level(logging.DEBUG)
-
-    protect_client._last_websocket_check = time.monotonic() - WEBSOCKET_CHECK_INTERVAL - 1
-    protect_client.disconnect_ws()
-    protect_client.ws_connection = None
-
-    active_ws = await protect_client.check_ws()
-
-    assert active_ws is True
-    expected_logs = ["Checking websocket", "Unifi OS: Websocket connection not active, failing back to polling"]
-    assert expected_logs == [rec.message for rec in caplog.records]
-    assert caplog.records[1].levelname == "WARNING"
+        assert mock_view is not None
+        assert len(view.keys()) == 2
+        assert view["id"] == mock_view["id"]
+        assert view["name"] == mock_view["name"]
